+++ integrations/gcp/gcp_core/search/iterators.py
@@ -1,7 +1,7 @@
 from typing import Any
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE
 from port_ocean.utils.async_iterators import stream_async_iterators_tasks
-from gcp_core.search.searches import search_all_projects
+from gcp_core.search.resource_searches import search_all_projects
 
 
 async def iterate_per_available_project(
