+++ integrations/gcp/gcp_core/search/resource_searches.py
